Fixes a compile warning * Follow up to 5ef29be. * Casts the size of the guild storage log to int16.

+++ src/map/clif.cpp
@@ -21020,7 +21020,7 @@ void clif_guild_storage_log( struct map_session_data* sd, std::vector<struct gui
 	int16 size = sizeof( struct PACKET_ZC_ACK_GUILDSTORAGE_LOG );
 
 	if( result == GUILDSTORAGE_LOG_FINAL_SUCCESS ){
-		size += log.size() * sizeof( struct PACKET_ZC_ACK_GUILDSTORAGE_LOG_sub );
+		size += static_cast<int16>(log.size()) * sizeof( struct PACKET_ZC_ACK_GUILDSTORAGE_LOG_sub );
 	}else{
 		log.clear();
 	}
